Split app-event-item-controls + stop logging expands for interacting with comments
