--- src/app/components/event-item/controls/controls.styl
@@ -1,4 +1,0 @@
-@require '~styles/variables'
-
-.-inline-comment
-	margin-top: ($grid-gutter-width / 2)
